Change BigDecimal to double

+++ src/main/java/com/polarbookshop/catalogservice/demo/BookDataLoader.java
@@ -7,7 +7,6 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Component
@@ -27,12 +26,12 @@ public class BookDataLoader {
         final var book1 = Book.of("1234567891",
                 "Northern Lights",
                 "Lyra Silverstar",
-                BigDecimal.valueOf(9.90),
+                9.90,
                 "Polarsophia");
         final var book2 = Book.of("1234567892",
                 "Polar Journey",
                 "Iorek Polarson",
-                BigDecimal.valueOf(12.90),
+                12.90,
                 "Polarsophia");
 
         bookRepository.saveAll(List.of(book1, book2));
